requirements.txt
fastapi
uvicorn
pydantic
#openai
pinecone
langchain
langchain-community
#langchain-openai
langchain-groq       
huggingface-hub
langchain-pinecone
PyPDF2
python-docx
python-dotenv
requests
gunicorn
langchain-huggingface
PyMuPDF